general final

+++ app/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Item(models.Model):
     name = models.CharField(max_length=200)
     description = models.CharField(max_length=250)
-    price = models.DecimalField(max_digits=6, decimal_places=2)
+    price = models.IntegerField()
 
     def __str__(self):
         return self.name
